Add Notes to the Unit struct
The Notes field holds some notes on the individual units and allows us to
record the background explanations of what the units are. This is important
as the world of non-metric units is littered with landmines and units with
identical names but different values.

The Unit definitions have all been enhanced to have (hopefully) useful
descriptions.